+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Changed the magic card detection. It now acts as flags (@iceman1001)
  - Added `hf ict` - skeleton commands in preperation (@iceman1001)
  - Added `lf hitag eview` - now supports viewing of emulator memory (@iceman1001)
  - Added `lf hitag view` - now supports viewing of dump files (@iceman1001)